Reorganize build to packages
