+++ build.gradle.kts
@@ -21,7 +21,7 @@ tasks.register<io.gitlab.arturbosch.detekt.report.ReportMergeTask>("detektReport
 	output.set(rootProject.buildDir.resolve("reports/detekt/merge.xml"))
 }
 
-tasks.register<net.twisterrob.sun.plugins.MergeLintSarifReportsTask>("lintReportMergeSarif") {
+tasks.register<net.twisterrob.sun.plugins.tasks.MergeLintSarifReportsTask>("lintReportMergeSarif") {
 	mergedSarifFile.set(rootProject.buildDir.resolve("reports/lint/merge.sarif"))
 }
