+++ gradle/plugins/src/main/kotlin/net/twisterrob/sun/plugins/internal/commonJavaConfig.kt
@@ -1,4 +1,4 @@
-package net.twisterrob.sun.plugins
+package net.twisterrob.sun.plugins.internal
 
 import io.gitlab.arturbosch.detekt.Detekt
 import io.gitlab.arturbosch.detekt.extensions.DetektExtension
